signature_distance: reorganize into categorized modules, share mnist_example checkpoint, fix Method B winner geometry
Consolidates ~20 scattered per-experiment .py files into streams.py,
distances.py, plots.py, adversarial_eval.py, models.py, and attacks.py,
switches the package to float64, and wires every model driver to load
mnist_example's shared trained SmallCNN/StrongCNN checkpoint instead of
training independent throwaway copies. That last change surfaced a real
bug: several adversarial-eval drivers previously trained a fresh 3-epoch
model each call, which gave both a spurious FGSM/PGD fold-ratio reversal
and an incorrect Method B hyperparameter-sweep winner (12h+4v). On the
canonical checkpoint, 16h+0v wins Stage B validation on both models and is
now the winner geometry used throughout.

Consolidates 7 notebooks into 2 (signatures_formation.ipynb,
adversarial_eval.ipynb) and 7 markdown docs into one README.md.

+++ signature_distance/tests/test_data_pool.py
@@ -15,7 +15,7 @@ def test_shapes_and_dtypes():
     images, labels = load_eval_pool(n_per_class=n_per_class, seed=0, root=str(DATA_ROOT))
     assert images.shape == (10 * n_per_class, 28, 28)
     assert labels.shape == (10 * n_per_class,)
-    assert images.dtype == torch.float32
+    assert images.dtype == torch.float64
     assert labels.dtype == torch.int64
